+++ packages/ts-newskit/src/components/slices/articleList/__tests__/index.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import '@testing-library/jest-dom';
-import { render, screen } from '../../../utils/test-utils';
+import { render, screen } from '../../../../utils/test-utils';
 import { ArticleListItem, ArticleListItemProps } from '../index';
 
 const renderComponent = (props: ArticleListItemProps) =>