+++ packages/article-skeleton/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.97.15](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.97.14...@times-components/article-skeleton@1.97.15) (2023-04-26)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
 ## [1.97.14](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.97.13...@times-components/article-skeleton@1.97.14) (2023-04-24)
 
 **Note:** Version bump only for package @times-components/article-skeleton