+++ packages/ssr/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.43.26](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.43.25...@times-components/ssr@2.43.26) (2023-04-26)
+
+**Note:** Version bump only for package @times-components/ssr
+
+
+
+
+
 ## [2.43.25](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.43.24...@times-components/ssr@2.43.25) (2023-04-24)
 
 **Note:** Version bump only for package @times-components/ssr